@@ -8,9 +8,10 @@
   "scripts": {
     "precommit": "lint-staged",
     "flow": "flow",
-    "lint": "eslint src",
-    "jest": "jest src",
-    "test": "npm run flow && npm run lint && npm run jest"
+    "lint": "eslint .", 
+    "jest": "jest", 
+    "test": "npm run flow && npm run lint && npm run jest", 
+    "format": "eslint --fix ."
   },
   "author": "Mike Grabowski <mike@callstack.io>",
   "repository": "github:callstack-io/haul",
@@ -65,24 +66,17 @@
   },
   "devDependencies": {
     "babel-eslint": "^7.2.1",
-    "eslint": "^3.18.0",
-    "eslint-config-airbnb": "^14.1.0",
-    "eslint-config-prettier": "^1.5.0",
-    "eslint-plugin-flowtype": "^2.30.4",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^4.0.0",
-    "eslint-plugin-react": "^6.10.3",
+    "eslint": "^4.3.0",
+    "eslint-config-callstack-io": "^0.3.1",
     "flow-bin": "^0.42.0",
     "husky": "^0.13.3",
     "jest": "^19.0.2",
     "lint-staged": "^3.4.0",
-    "prettier": "^0.22.0",
     "react": "16.0.0-alpha.12"
   },
   "lint-staged": {
     "*.js": [
       "eslint --fix",
-      "prettier --single-quote --trailing-comma all --write",
       "git add"
     ]
   },
@@ -104,42 +98,33 @@
     ]
   },
   "eslintConfig": {
-    "parser": "babel-eslint",
-    "extends": [
-      "airbnb",
-      "prettier",
-      "prettier/flowtype",
-      "prettier/react"
-    ],
-    "env": {
-      "node": true,
-      "es6": true
-    },
-    "parserOptions": {
-      "sourceType": "module"
-    },
+    "extends": "callstack-io",
     "rules": {
       "strict": 0,
       "arrow-body-style": 0,
       "arrow-parens": 0,
       "class-methods-use-this": 0,
-      "default-case": 0,
-      "generator-star-spacing": 0,
+      "flowtype/no-weak-types": 0,
+      "flowtype/require-parameter-type": 0,
       "global-require": 0,
-      "new-cap": 0,
       "no-alert": 0,
       "no-confusing-arrow": 0,
       "no-console": 0,
       "no-constant-condition": 0,
-      "no-duplicate-imports": 0,
-      "no-plusplus": 0,
-      "no-restricted-syntax": 0,
-      "no-underscore-dangle": 0,
-      "no-use-before-define": 0,
-      "import/no-duplicates": 0,
-      "import/extensions": 0,
-      "import/no-unresolved": 0,
-      "import/no-dynamic-require": 0
+      "no-restricted-syntax": 0
+    },
+    "overrides": {
+      "files": "**/integration_tests/**/*.js",
+      "rules": {
+        "import/no-extraneous-dependencies": [
+          "error",
+          {
+            "devDependencies": [
+              "**/integration_tests/**"
+            ]
+          }
+        ]
+      }
     }
   }
 }
